Add alias field and property to ContactInfo class
A new private field `alias` of type `string` has been added to the
`ContactInfo` class, initialized to an empty string. A corresponding
public property `Alias` has been introduced, which includes a getter
and setter for the `alias` field. This property is decorated with the
`DefaultValueStringEmpty` attribute and includes a summary comment
explaining its purpose as a user-set friendly name.